I was able to add the generated Texture2D to the project's assets and apply it to the light that it was generated for. I was hopping to create an exr image texture, add it to the project, alter it's import settings to suit our needs, and then apply the result of that to the light source we generated the image for; yet, that didn't work. I was able to alter the settings if the image was there from a previous run, but I ran into issues when it came to applying the settings to an image that was generated on the same run as which we tried to apply those settings.
This will create an asset that's as small as possible with zero compression, but it won't allow the
developers using the tool to change the import settings to something that better suits their needs.
It will also make it more difficult for them to import the image into an outside program and make
alterations or additions of their own.