Downgraded to Unity 2018.4.4f1 because I was having some UI freezing issues. It looked like all the background processes were still running, but at some point in the preparation stage the UI freezes, and this then causes a few other things to not work. I tried it a few times with both versions, it froze with both versions. That said, the 1024x1024 resolution hasn't frozen it for 2018.4.4f1 and it would sometimes freeze it for 1024x1024. 2048x2048 keeps freezing it for both.
I found that the reason why I had to do that odd uv flip was because I was getting bad half-size values. And, the reason for the bad half-size value was because the tan() function in the shader was expecting a radian value and I was giving it a value that was in degrees. I was able to get a better look at this once I was able to try out more spot-light deg. values.